Stable version of Java wrapper and JHove for the lastest revision of warc-tools

+++ trunk/contrib/java/src/wfile.java
@@ -1,5 +1,7 @@
 package warctools;
 
+
+
 import com.sun.jna.Pointer;
 
 
@@ -9,12 +11,12 @@ public class wfile implements wfilei {
 	public Pointer file = null;
 	
 	
-public wfile (String path, int dim, int mode, int compressed, String dname)
+public wfile (String path, long dim, int mode, int compressed, String dname)
 {
 	file = bless_Wfile( path,  dim,  mode,  compressed,  dname);
 }
 
-public Pointer bless_Wfile (String path, int dim, int mode, int compressed, String dname )
+public Pointer bless_Wfile (String path, long dim, int mode, int compressed, String dname )
 {
 	return wfilei.INSTANCE.bless_Wfile( path,  dim,  mode,  compressed,  dname);
 }
@@ -29,7 +31,7 @@ public Pointer WFile_nextRecord     (Pointer x)
 	return wfilei.INSTANCE.WFile_nextRecord(x);
 }
 
-public long      WFile_isFull          (Pointer x )
+public int      WFile_isFull          (Pointer x )
 {
 	return wfilei.INSTANCE.WFile_isFull(x);
 }
@@ -49,7 +51,7 @@ public int     WFile_seek           (Pointer x,  long offset)
 	return wfilei.INSTANCE.WFile_seek(x,offset);
 }
 
-public void     destroy             (Pointer x)
+public void     destroy           (Pointer x)
 {
 	wfilei.INSTANCE.destroy(x);
 }
